+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/OneTapFlow+Screens.swift
@@ -40,13 +40,12 @@ extension OneTapFlow {
             // WARNING: Keep strong ref here (or any other block for this initializer) or it'll release the object after creating it
             self.executeNextStep()
         }
-        let viewModel = model.reviewConfirmViewModel()
+        let viewModel = model.oneTapViewModel()
+        let reviewVC = PXOneTapViewController(viewModel: viewModel, timeOutPayButton: model.getTimeoutForOneTapReviewController(), callbackPaymentData: callbackPaymentData, callbackConfirm: callbackConfirm, callbackUpdatePaymentOption: callbackUpdatePaymentOption, callbackExit: callbackExit, finishButtonAnimation: finishButtonAnimation)
         
-        let reviewVC = PXOneTapViewController(viewModel: viewModel, navigationHandler: pxNavigationHandler, timeOutPayButton: model.getTimeoutForOneTapReviewController(), callbackPaymentData: callbackPaymentData, callbackConfirm: callbackConfirm, callbackUpdatePaymentOption: callbackUpdatePaymentOption, callbackExit: callbackExit, finishButtonAnimation: finishButtonAnimation)
-
         pxNavigationHandler.pushViewController(viewController: reviewVC, animated: true)
     }
-
+    
     func showSecurityCodeScreen() {
         let securityCodeVc = SecurityCodeViewController(viewModel: model.savedCardSecurityCodeViewModel(), collectSecurityCodeCallback: { [weak self] (_, securityCode: String) -> Void in
             self?.getTokenizationService().createCardToken(securityCode: securityCode)