+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapViewController.swift
@@ -14,7 +14,6 @@ final class PXOneTapViewController: PXComponentContainerViewController {
     // MARK: Definitions
     lazy var itemViews = [UIView]()
     fileprivate var viewModel: PXOneTapViewModel
-    fileprivate var navigationHandler: PXNavigationHandler
     private var discountTermsConditionView: PXTermsAndConditionView?
 
     let slider = PXCardSlider()
@@ -39,9 +38,8 @@ final class PXOneTapViewController: PXComponentContainerViewController {
     private var navigationBarTapGesture: UITapGestureRecognizer?
 
     // MARK: Lifecycle/Publics
-    init(viewModel: PXOneTapViewModel, navigationHandler: PXNavigationHandler, timeOutPayButton: TimeInterval = 15, callbackPaymentData : @escaping ((PXPaymentData) -> Void), callbackConfirm: @escaping ((PXPaymentData, Bool) -> Void), callbackUpdatePaymentOption: @escaping ((PaymentMethodOption) -> Void), callbackExit: @escaping (() -> Void), finishButtonAnimation: @escaping (() -> Void)) {
+    init(viewModel: PXOneTapViewModel, timeOutPayButton: TimeInterval = 15, callbackPaymentData : @escaping ((PXPaymentData) -> Void), callbackConfirm: @escaping ((PXPaymentData, Bool) -> Void), callbackUpdatePaymentOption: @escaping ((PaymentMethodOption) -> Void), callbackExit: @escaping (() -> Void), finishButtonAnimation: @escaping (() -> Void)) {
         self.viewModel = viewModel
-        self.navigationHandler = navigationHandler
         self.callbackPaymentData = callbackPaymentData
         self.callbackConfirm = callbackConfirm
         self.callbackExit = callbackExit
@@ -163,7 +161,7 @@ extension PXOneTapViewController {
             PXLayout.pinBottom(view: footerView, withMargin: bottomMargin).isActive = true
         }
 
-        if let selectedCard = selectedCard, selectedCard.isDisabled {
+        if let selectedCard = selectedCard, !selectedCard.status.enabled {
             loadingButtonComponent?.setDisabled()
         }
 
@@ -258,7 +256,7 @@ extension PXOneTapViewController {
     }
 
     private func confirmPayment() {
-        if viewModel.shouldValidateWithBiometric(withCardId: selectedCard?.cardId) {
+        if viewModel.shouldValidateWithBiometric() {
             let biometricModule = PXConfiguratorManager.biometricProtocol
             biometricModule.validate(config: PXConfiguratorManager.biometricConfig, onSuccess: { [weak self] in
                 DispatchQueue.main.async {
@@ -386,9 +384,10 @@ extension PXOneTapViewController: PXCardSliderProtocol {
 
         // Add card. - card o credits payment method selected
         let validData = targetModel.cardData != nil || targetModel.isCredits
-        let shouldDisplay = validData && !targetModel.isDisabled
+        let shouldDisplay = validData && targetModel.status.enabled
         if shouldDisplay {
             displayCard(targetModel: targetModel)
+            loadingButtonComponent?.setEnabled()
         } else {
             loadingButtonComponent?.setDisabled()
             headerView?.updateModel(viewModel.getHeaderViewModel(selectedCard: nil))
@@ -421,9 +420,23 @@ extension PXOneTapViewController: PXCardSliderProtocol {
         }
     }
 
-    func disabledCardDidTap(isAccountMoney: Bool) {
-        let vc = PXDisabledViewController(isAccountMoney: isAccountMoney)
-        PXComponentFactory.Modal.show(viewController: vc, title: nil)
+    func disabledCardDidTap(status: PXStatus) {
+        showDisabledCardModal(status: status)
+    }
+
+    func showDisabledCardModal(status: PXStatus) {
+        guard let message = status.secondaryMessage?.message else {return}
+        let vc = PXOneTapDisabledViewController(text: message)
+        let buttonTitle = "px_dialog_detail_payment_method_disable_link".localized
+        PXComponentFactory.Modal.show(viewController: vc, title: nil, actionTitle: buttonTitle, actionBlock: {
+            //Select first item
+            self.slider.goToItemAt(index: 0, animated: false)
+            if let card = self.viewModel.getCardSliderViewModel().first {
+                self.newCardDidSelected(targetModel: card)
+            }
+        })
+
+        trackScreen(path: TrackingPaths.Screens.OneTap.getOneTapDisabledModalPath(), treatAsViewController: false)
     }
 
     func addPaymentMethodCardDidTap() {
@@ -439,12 +452,6 @@ extension PXOneTapViewController: PXCardSliderProtocol {
 
             let cardFormVC = MLCardForm(builder: builder).setupController()
             navigationController?.pushViewController(cardFormVC, animated: true)
-
-//            let cardFormController = NewCardAssociationViewController(model: "modelo de prueba")
-//            cardFormController.delegate = self
-//            navigationHandler.pushViewController(cleanCompletedCheckouts: false, targetVC: cardFormController,
-//                                                 animated: true)
-            //navigationController?.pushViewController(cardFormController, animated: true)
         }
     }
 
@@ -459,13 +466,13 @@ extension PXOneTapViewController: PXCardSliderProtocol {
 
 extension PXOneTapViewController: MLCardFormAddCardProtocol {
     internal func didAddCard(cardInfo: [String: String]) {
-        let pxCardSliderViewModel = createNewMockedCard()
-        var cardSliderViewModel = viewModel.getCardSliderViewModel()
-        cardSliderViewModel.insert(pxCardSliderViewModel, at: cardSliderViewModel.count - 1)
-        slider.update(cardSliderViewModel)
-        viewModel.updateCardSliderViewModel(pxCardSliderViewModel: cardSliderViewModel)
-        newCardDidSelected(targetModel: pxCardSliderViewModel)
-        installmentInfoRow?.model = viewModel.getInstallmentInfoViewModel()
+//        let pxCardSliderViewModel = createNewMockedCard()
+//        var cardSliderViewModel = viewModel.getCardSliderViewModel()
+//        cardSliderViewModel.insert(pxCardSliderViewModel, at: cardSliderViewModel.count - 1)
+//        slider.update(cardSliderViewModel)
+//        viewModel.updateCardSliderViewModel(pxCardSliderViewModel: cardSliderViewModel)
+//        newCardDidSelected(targetModel: pxCardSliderViewModel)
+//        installmentInfoRow?.model = viewModel.getInstallmentInfoViewModel()
         navigationController?.popViewController(animated: true)
     }
 
@@ -474,48 +481,52 @@ extension PXOneTapViewController: MLCardFormAddCardProtocol {
     }
 }
 
-// MARK: Mocked Data
-private extension PXOneTapViewController {
-    func createNewMockedCard() -> PXCardSliderViewModel {
-        let payerCost = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 0, maxAllowedAmount: 60000, recommendedMessage: "1 Parcela de R$ 100", installmentAmount: 100, totalAmount: 100, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil)
-        var payerCostArray = [PXPayerCost]()
-        payerCostArray.append(payerCost)
-        /************ SPLIT CONFIGURATION MOCKED ***********/
-        let splitConfig = getMockedSplitConfiguration()
-        /*******************************/
-        let amountConfig = PXAmountConfiguration(selectedPayerCostIndex: 0, payerCosts: payerCostArray, splitConfiguration: splitConfig, discountToken: nil, amount: nil)
-        let cardData = PXCardDataFactory().create(cardName: "APRO BADO", cardNumber: "************5682", cardCode: "", cardExpiration: "2/24")
-        let mockedCard = PXCardSliderViewModel("visa", "credit_card", "25", TemplateCard(), cardData, payerCostArray, payerCost, "8755873036", true, amountConfiguration: amountConfig, creditsViewModel: nil, isDisabled: false)
-        return mockedCard
-    }
-
-    func getMockedSplitConfiguration() -> PXSplitConfiguration {
-        var splitPayerCostArray = [PXPayerCost]()
-        let firstPayerCost = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "1 cuota de $ 2342.,47 ($ 2.342,47)", installmentAmount: 2342.4699999, totalAmount: 2342.4699999, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil, agreements: [PXAgreement]())
-        splitPayerCostArray.append(firstPayerCost)
-
-        let primarySplitPaymentMethod = PXSplitPaymentMethod(amount: 2342.469999, id: "", discount: nil, message: nil, selectedPayerCostIndex: 0, payerCosts: splitPayerCostArray)
-        let secondarySplitPaymentMethod = PXSplitPaymentMethod(amount: 1157.53, id: "account_money", discount: nil, message: "en Mercado Pago", selectedPayerCostIndex: nil, payerCosts: nil)
-
-        let splitConfiguration = PXSplitConfiguration(primaryPaymentMethod: primarySplitPaymentMethod, secondaryPaymentMethod: secondarySplitPaymentMethod, splitEnabled: false)
-        return splitConfiguration
-    }
-
-    func createNewMockedCard2() -> PXCardSliderViewModel {
-        let payerCost1 = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "1 cuota de 100", installmentAmount: 100, totalAmount: 100, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil)
-        //        let payerCost2 = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "2 cuota de 200", installmentAmount: 100, totalAmount: 100, installments: 2, processingMode: "aggregator", paymentMethodOptionId: nil)
-        var payerCostArray = [PXPayerCost]()
-        payerCostArray.append(payerCost1)
-        //        payerCostArray.append(payerCost2)
-        let amountConfig = PXAmountConfiguration(selectedPayerCostIndex: 0, payerCosts: payerCostArray, splitConfiguration: nil, discountToken: nil, amount: nil)
-        let cardData = PXCardDataFactory().create(cardName: "ESTEBAN BOFFA", cardNumber: "************1752", cardCode: "", cardExpiration: "10/20")
-        let mockedCard = PXCardSliderViewModel("visa", "credit_card", "310", TemplateCard(), cardData, payerCostArray, payerCost1, "8735161676", true, amountConfiguration: amountConfig, creditsViewModel: nil, isDisabled: false)
-        return mockedCard
-    }
-}
+//// MARK: Mocked Data
+//private extension PXOneTapViewController {
+//    func createNewMockedCard() -> PXCardSliderViewModel {
+//        let payerCost = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 0, maxAllowedAmount: 60000, recommendedMessage: "1 Parcela de R$ 100", installmentAmount: 100, totalAmount: 100, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil)
+//        var payerCostArray = [PXPayerCost]()
+//        payerCostArray.append(payerCost)
+//        /************ SPLIT CONFIGURATION MOCKED ***********/
+//        let splitConfig = getMockedSplitConfiguration()
+//        /*******************************/
+//        let amountConfig = PXAmountConfiguration(selectedPayerCostIndex: 0, payerCosts: payerCostArray, splitConfiguration: splitConfig, discountToken: nil, amount: nil)
+//        let cardData = PXCardDataFactory().create(cardName: "APRO BADO", cardNumber: "************5682", cardCode: "", cardExpiration: "2/24")
+//        let mockedCard = PXCardSliderViewModel("visa", "credit_card", "25", TemplateCard(), cardData, payerCostArray, payerCost, "8755873036", true, amountConfiguration: amountConfig, creditsViewModel: nil, isDisabled: false)
+//        return mockedCard
+//    }
+//
+//    func getMockedSplitConfiguration() -> PXSplitConfiguration {
+//        var splitPayerCostArray = [PXPayerCost]()
+//        let firstPayerCost = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "1 cuota de $ 2342.,47 ($ 2.342,47)", installmentAmount: 2342.4699999, totalAmount: 2342.4699999, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil, agreements: [PXAgreement]())
+//        splitPayerCostArray.append(firstPayerCost)
+//
+//        let primarySplitPaymentMethod = PXSplitPaymentMethod(amount: 2342.469999, id: "", discount: nil, message: nil, selectedPayerCostIndex: 0, payerCosts: splitPayerCostArray)
+//        let secondarySplitPaymentMethod = PXSplitPaymentMethod(amount: 1157.53, id: "account_money", discount: nil, message: "en Mercado Pago", selectedPayerCostIndex: nil, payerCosts: nil)
+//
+//        let splitConfiguration = PXSplitConfiguration(primaryPaymentMethod: primarySplitPaymentMethod, secondaryPaymentMethod: secondarySplitPaymentMethod, splitEnabled: false)
+//        return splitConfiguration
+//    }
+//
+//    func createNewMockedCard2() -> PXCardSliderViewModel {
+//        let payerCost1 = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "1 cuota de 100", installmentAmount: 100, totalAmount: 100, installments: 1, processingMode: "aggregator", paymentMethodOptionId: nil)
+//        //        let payerCost2 = PXPayerCost(installmentRate: 0, labels: [String](), minAllowedAmount: 1, maxAllowedAmount: 700000, recommendedMessage: "2 cuota de 200", installmentAmount: 100, totalAmount: 100, installments: 2, processingMode: "aggregator", paymentMethodOptionId: nil)
+//        var payerCostArray = [PXPayerCost]()
+//        payerCostArray.append(payerCost1)
+//        //        payerCostArray.append(payerCost2)
+//        let amountConfig = PXAmountConfiguration(selectedPayerCostIndex: 0, payerCosts: payerCostArray, splitConfiguration: nil, discountToken: nil, amount: nil)
+//        let cardData = PXCardDataFactory().create(cardName: "ESTEBAN BOFFA", cardNumber: "************1752", cardCode: "", cardExpiration: "10/20")
+//        let mockedCard = PXCardSliderViewModel("visa", "credit_card", "310", TemplateCard(), cardData, payerCostArray, payerCost1, "8735161676", true, amountConfiguration: amountConfig, creditsViewModel: nil, isDisabled: false)
+//        return mockedCard
+//    }
+//}
 
 // MARK: Installment Row Info delegate.
 extension PXOneTapViewController: PXOneTapInstallmentInfoViewProtocol, PXOneTapInstallmentsSelectorProtocol {
+    func disabledCardTapped(status: PXStatus) {
+        showDisabledCardModal(status: status)
+    }
+
     func payerCostSelected(_ payerCost: PXPayerCost) {
         // Update cardSliderViewModel
         if let infoRow = installmentInfoRow, viewModel.updateCardSliderViewModel(newPayerCost: payerCost, forIndex: infoRow.getActiveRowIndex()) {