+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXCardSlider/PXCardSliderPagerCell.swift
@@ -15,7 +15,6 @@ class PXCardSliderPagerCell: FSPagerViewCell {
 
     private lazy var cornerRadius: CGFloat = 11
     private var cardHeader: MLCardDrawerController?
-    private var warningBadgeIcon: UIView!
 
     @IBOutlet weak var containerView: UIView!
 
@@ -32,7 +31,7 @@ class PXCardSliderPagerCell: FSPagerViewCell {
 
 // MARK: Publics.
 extension PXCardSliderPagerCell {
-    func render(withCard: CardUI, cardData: CardData, isDisabled: Bool, cardSize: CGSize) {
+    func render(withCard: CardUI, cardData: CardData, isDisabled: Bool, cardSize: CGSize, bottomMessage: String? = nil) {
         containerView.layer.masksToBounds = false
         containerView.removeAllSubviews()
         containerView.layer.cornerRadius = cornerRadius
@@ -47,27 +46,28 @@ extension PXCardSliderPagerCell {
             PXLayout.centerHorizontally(view: headerView).isActive = true
             PXLayout.centerVertically(view: headerView).isActive = true
         }
-        addWarningBadge(isDisabled)
+        addBottomMessageView(message: bottomMessage)
     }
 
-    func renderEmptyCard(cardSize: CGSize) {
+    func renderEmptyCard(title: PXText? = nil, cardSize: CGSize) {
         containerView.layer.masksToBounds = false
         containerView.removeAllSubviews()
         containerView.layer.cornerRadius = cornerRadius
         containerView.backgroundColor = .clear
-        cardHeader = MLCardDrawerController(EmptyCard(), PXCardDataFactory(), false)
+        let emptyCard = EmptyCard(title: title)
+        cardHeader = MLCardDrawerController(emptyCard, PXCardDataFactory(), false)
         cardHeader?.view.frame = CGRect(origin: CGPoint.zero, size: cardSize)
         cardHeader?.animated(false)
         cardHeader?.show()
         if let headerView = cardHeader?.view {
             containerView.addSubview(headerView)
-            EmptyCard.render(containerView: containerView)
+            emptyCard.render(containerView: containerView)
             PXLayout.centerHorizontally(view: headerView).isActive = true
             PXLayout.centerVertically(view: headerView).isActive = true
         }
     }
 
-    func renderAccountMoneyCard(balanceText: String, isDisabled: Bool, cardSize: CGSize) {
+    func renderAccountMoneyCard(isDisabled: Bool, cardSize: CGSize, bottomMessage: String? = nil) {
         containerView.layer.masksToBounds = false
         containerView.backgroundColor = .clear
         containerView.removeAllSubviews()
@@ -79,16 +79,15 @@ extension PXCardSliderPagerCell {
 
         if let headerView = cardHeader?.view {
             containerView.addSubview(headerView)
-            AccountMoneyCard.render(containerView: containerView, balanceText: balanceText, isDisabled: isDisabled, size: cardSize)
+            AccountMoneyCard.render(containerView: containerView, isDisabled: isDisabled, size: cardSize)
             PXLayout.centerHorizontally(view: headerView).isActive = true
             PXLayout.centerVertically(view: headerView).isActive = true
         }
-        addWarningBadge(isDisabled)
+        addBottomMessageView(message: bottomMessage)
     }
 
-
-    func renderConsumerCreditsCard(creditsViewModel: PXCreditsViewModel, isDisabled: Bool, cardSize: CGSize) {
-        consumerCreditCard = ConsumerCreditsCard(creditsViewModel)
+    func renderConsumerCreditsCard(creditsViewModel: PXCreditsViewModel, isDisabled: Bool, cardSize: CGSize, bottomMessage: String? = nil) {
+        consumerCreditCard = ConsumerCreditsCard(creditsViewModel, isDisabled: isDisabled)
         guard let consumerCreditCard = consumerCreditCard else { return }
 
         containerView.layer.masksToBounds = false
@@ -109,19 +108,43 @@ extension PXCardSliderPagerCell {
             PXLayout.centerHorizontally(view: headerView).isActive = true
             PXLayout.centerVertically(view: headerView).isActive = true
         }
-        addWarningBadge(isDisabled)
+        addBottomMessageView(message: bottomMessage)
     }
 
-    func addWarningBadge(_ isDisabled: Bool) {
-        if isDisabled {
-            let image = ResourceManager.shared.getImage("warning_badge")
-            warningBadgeIcon = UIImageView(image: image)
-            containerView.insertSubview(warningBadgeIcon, at: 10)
-            PXLayout.setHeight(owner: warningBadgeIcon, height: 60).isActive = true
-            PXLayout.setWidth(owner: warningBadgeIcon, width: 60).isActive = true
-            PXLayout.pinTop(view: warningBadgeIcon, withMargin: -28).isActive = true
-            PXLayout.pinRight(view: warningBadgeIcon, withMargin: PXLayout.S_MARGIN).isActive = true
-        }
+    func addBottomMessageView(message: String?) {
+        guard let message = message else {return}
+        let messageView = UIView()
+        messageView.translatesAutoresizingMaskIntoConstraints = false
+        messageView.backgroundColor = ThemeManager.shared.noTaxAndDiscountLabelTintColor()
+
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = message
+        label.numberOfLines = 1
+        label.textAlignment = .center
+        label.textColor = .white
+        label.font = Utils.getSemiBoldFont(size: PXLayout.XXXS_FONT)
+
+        messageView.addSubview(label)
+
+        NSLayoutConstraint.activate([
+            label.leadingAnchor.constraint(equalTo: messageView.leadingAnchor),
+            label.trailingAnchor.constraint(equalTo: messageView.trailingAnchor),
+            label.topAnchor.constraint(equalTo: messageView.topAnchor),
+            label.bottomAnchor.constraint(equalTo: messageView.bottomAnchor)
+            ])
+
+        self.containerView.clipsToBounds = true
+        self.containerView.addSubview(messageView)
+
+        NSLayoutConstraint.activate([
+            messageView.leadingAnchor.constraint(equalTo: self.containerView.leadingAnchor),
+            messageView.trailingAnchor.constraint(equalTo: self.containerView.trailingAnchor),
+            messageView.heightAnchor.constraint(equalToConstant: 24),
+            messageView.bottomAnchor.constraint(equalTo: self.containerView.bottomAnchor)
+            ])
+
+        self.layoutIfNeeded()
     }
 
     func flipToBack() {